Scrolling works, this fixes a flat allocation bounds issue by using the first fit allocater instead.

+++ src/kernel/memory/first-fit/block.h
@@ -9,18 +9,18 @@ namespace kernel {
       class __attribute__((packed)) Block {
         bool allocatedp : 1;            /// True when block is allocated.
         size_t start : 23;               /// Pointer to start of block.
-        uint16_t size : 16;              /// Size in bytes of the block.
+        size_t size : 16;              /// Size in bytes of the block.
 
 
       public:
         Block() : allocatedp(false), start(0x0), size(0x00) {};
         /// Mark block as used.
-        void* alloc(size_t start, uint16_t size);
+        void* alloc(size_t start, size_t size);
         void free(void);            /// Mark block as free.
 
         /// Conditionally free \ref memory_address.
         bool freeIf(size_t memory_address);
-        void* allocIf(uint16_t size);
+        void* allocIf(size_t size);
         bool brandNewP(void);
 
         void inspect (uint8_t line) {
